Jenkins/core-backup/requirements.txt
requests
py_dotenv
aiohttp==3.8.2
yarl==1.8.2
frozenlist==1.3.1
PyGithub

